requirements.txt
cohere
python-dotenv
chromadb
fastapi
uvicorn
requests
beautifulsoup4